refactor: move interfaces in separate file
