+++ src/game/index.ts
@@ -1,35 +1,4 @@
-interface IShipData {
-  position: {
-    x: number;
-    y: number;
-  };
-  direction: boolean;
-  length: number;
-  type: 'small' | 'medium' | 'large' | 'huge';
-}
-
-interface IShipsPlayer {
-  playerIndex: number;
-  ships: IShipData[];
-}
-
-interface IField {
-  playerIndex: number;
-  field: string[][];
-}
-
-interface IShipNumber {
-  playerIndex: number;
-  ships: number;
-}
-
-enum InfoField {
-  None = 'none',
-  Healthy = 'healthy',
-  Hit = 'hit',
-  Miss = 'miss',
-  Killed = 'killed',
-}
+import { IShipData, IShipsPlayer, IField, IShipNumber, InfoField } from '../interfaces/game';
 
 export class BattleshipGame {
   static id = 0;
